requirements.txt
click>=8.0
rich>=13.0
requests>=2.28
PyYAML>=6.0
pytest>=7.0
